Simplified Packet Handler Registration (#54)
Now that we have the "[PacketID()]" attribute we can apply the same principles
to registering packet handlers. A new "[PacketHandler]" attribute will wire up
the handler to DI and the PacketProcessor automatically. There are also some
unit tests to prevent developers from accidentally binding handlers incorrectly.

+++ fom-network/include/fom-network/FOMDataSerializer.h
@@ -32,7 +32,8 @@ class FOM_API FOMDataSerializer {
    */
   static const bool ShouldIgnorePacket(PacketIdentifier id) {
     // Never ignore game packets.
-    if (id >= ID_USER_PACKET_ENUM) return false;
+    if (static_cast<uint8_t>(id) >= static_cast<uint8_t>(ID_USER_PACKET_ENUM))
+      return false;
 
     // RakNet client packets are "serialized" but we don't actually need to do
     // anything with the data since we're only forwarding the ID to the
